develop: moved ChangeableList.java from objects to executors
